backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
python-dotenv==1.0.1
pydantic==2.8.2
python-multipart==0.0.9
boto3==1.35.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
